+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.22.0] - 2023-11-08
+
 ### Changed
 
 - Removed duplicated tags before creating the OIDC provider.
@@ -311,7 +313,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-03-04
 
-[Unreleased]: https://github.com/giantswarm/irsa-operator/compare/v0.21.0...HEAD
+[Unreleased]: https://github.com/giantswarm/irsa-operator/compare/v0.22.0...HEAD
+[0.22.0]: https://github.com/giantswarm/irsa-operator/compare/v0.21.0...v0.22.0
 [0.21.0]: https://github.com/giantswarm/irsa-operator/compare/v0.20.0...v0.21.0
 [0.20.0]: https://github.com/giantswarm/irsa-operator/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/giantswarm/irsa-operator/compare/v0.18.0...v0.19.0